Рефакторинг: разбивка на файлы + удаление неиспользуемого кода
